refactor: MathpixApiPdfProcessClient로 역할 분리
